curl: Backport CVE-2016-5420 fix from cURL 7.50.1
Bug: libcurl supports reuse of established connections for subsequent
requests. It does this by keeping a few previous connections "alive" in
a connection pool so that a subsequent request that can use one of them
instead of creating a new connection will do so.

When using a client certificate for a connection that was then put into
the connection pool, that connection could then wrongly get reused in a
subsequent request to that same server that either didn't use a client
certificate at all or that asked to use a different client certificate
thus trying to tell the user that it is a different entity.

Backported fix: Only reuse connections with the same client cert.

https://curl.haxx.se/docs/adv_20160803B.html

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-CAR-ID: 599706
Natinst-ReviewBoard-ID: 157131