niacctbase: Move static UIDs and GIDs
With recent changes to useradd-staticids.bbclass in krogoth, static IDs must
be set via the passwd and group files. Previously these users could be added
via the USERADD_PARAM as long as a static ID was listed.

Remove the static IDs from the USERADD_PARAM variable in niacctbase and add
them to passwd and group files.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>