ni-external-components: clean up and enable nirtcfg on arm
Migration to newer nilrt OS on ARM leaves the target with networking down
because there was no nirtcfg hardfloat ipk to install during migration dev.

After nirtcfg was compiled for hardfloat we postponed adding it to the base
image because it will go away (replaced by connman), but because the
connman work was un-prioritiezed we continue to have ARM targets which get
migrated with networking down, so we add it to ARM to fix migration so
migration users can use ssh & other tools without needing a serial console
just to enable networking.

Once connman will be made the default network manager on newer nilrt on ARM
we can simply remove nirtcfg, but until then I think it's more important to
fix migration to have networking.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>