xserver-xorg: Add udev rule for display hotplug
When unplugging a display that is using X, the system will detect
the re-plug of the monitor but not jump to any hasty conclusions
(like that you want to use that thing that you just plugged in).

Create a udev rule that will catch a monitor hotplug and tell X
to use that monitor.

Create the xserver-xorg-udev-rules package to contain that udev
rule. Set the RDEPENDS and RRECOMMENDS appropriately.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Natinst-ReviewBoard-ID: 165151
Natinst-CAR-ID: 613055