salt: make RDEPENDS on python version agnostic
Currently, we install python via python-zip, which Provides python. The
salt packages have version restriction on python, which is an issue
since Debian policy says:

"If a relationship field has a version number attached, only real
packages will be considered to see whether the relationship is satisfied
(or the prohibition violated, for a conflict or breakage).

Opkg 0.3.3 and below don't enforces that rule. But opkg 0.3.4 does.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>