salt: add avahi-ui to DEPENDS
NILRT needs the salt-minion package provided by salt to depend on
the python-avahi package provided by the avahi-ui recipe. This is done
in the bbappend.

However because avahi-ui is not marked as a dependency for salt, a build
race is triggered. On some setups, depending on the build order, we get:

ERROR: Nothing RPROVIDES 'python-avahi' (but (...)/meta-cloud-services/meta-openstack/recipes-support/salt/salt_2016.3.0.bb RDEPENDS on or otherwise requires it)

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>