Revert "salt: add python-avahi to RDEPENDS_-minion"
This reverts commit c2388587d44fd597d36c62aee4e7e0b52159bdb8.

python-avahi could not be built on NILRT 17.0 in the extra feed because
it was contained in the avahi-ui bb recipe, which depended on x11 being in
DISTRO_FEATURES. The extra feed is built using the "nilrt.conf" distro and
not the "nilrt-xfce.conf" (only this has x11 in DISTRO_FEATURES). Building
the extra feed with nilrt-xfce.conf is not an option so we can't actually
build avahi-ui in the 17.0 extra feeds.

On top of this, OE-core upstream post-krogoth removed python-avahi from
avahi-ui completely because no-one was using it in commit 36e9ed899506f
("avahi-ui: remove support for building a python module").

So we have to find a replacement, probably write a new recipe, maybe
python-zeroconf?

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>