grub: add grub-native package containing grub-mkimage
A native grub-mkimage is needed by wic to create a boot iso on the host
build machine. If we don't provide grub-native, wic errors with:

Error: A native program grub-mkimage required to build the image was not found (see details above).

Wic failed to find a recipe to build native grub-mkimage. Please file a bug against wic.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>