grub: split and simplify recipe bbappend
Turns out we don't have to upstream/refactor anything in OE-core, because
there is a grub-efi_git.bb in meta-measured which does exactly what we
need (there is a some logic duplication between meta-measured/OE-core but
it's no problem for meta-nilrt).

So now split all efi logic and move it to a bbappend of its own based on
the meta-measured recipe. Also move common logic in grub-nilrt.inc the
same way we do with the kernel based recipe which uses linux-nilrt.inc.

Maybe we should move the grub-efi_git.bb recipe from meta-measured to
OE-core because it's more maintained and it could eliminate the duplicate
logic between them? Anyway, that change does not impact this.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>