niacctbase: add openvpn user to network group
We need this to fix openvpn accessing /sbin/ip which will be also
added to the network group.

On the older nilrt OS, groups and users ID's are statically asigned
using the files/group files/users files inside this repo, so adding
them here will use the predefined values in those files.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>