iproute2_4.%.bbappend: install 'ip' using network group
Openvpn needs to access this tool, so chown it to the network group of
which the openvpn user is also part of. We do this to allow the
privilege-dropped openvpn daemon to configure the network (note that
ip is now setuid).

No other user/program runs in the network group at this time.

The symlink created by update-alternatives is owned by root:root but
it does not matter because on Linux symlink permissions are ignored,
only the target permissions matter to be set correctly.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>