udev-extraconf: Remove broken ifrename rules for CVS-1458RT, IC-317x
These rules stopped working in krogoth because of the upgrade to eudev,
and more specifically, eudev commit
d9488d78ebc2dde8ac4065381780e4d974056e92 ("rule-generator: proactively
rename ifaces to avoid conflicts"), which more or less broke temporary
renames unless the rules generator is enabled. (And it's not for OE.)

The udev old-style ifrename subsystem is a crufty mess, and it would be
beneficial to avoid it. Furthermore, these rules are implicitly
controller-specific; even though they were carefully crafted to apply to
these specific controllers, this is conceptually not the right approach
to applying this policy.

Natinst-CAR-ID: 617482
Natinst-ReviewBoard-ID: 169321
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>