bash_4.%.bbappend: remove NON_INTERACTIVE_LOGIN_SHELLS cflag
This effectively disables sourcing /etc/profile in non-interactive
bash shells because it crashes embedded UI apps like thunar. This
happens because we set hard ulimits for stack sizes to 256kb in
/etc/profile.

OE-core commit 33af68d62bb427 ("bash: explicitly define
NON_INTERACTIVE_LOGIN_SHELLS in CFLAGS") changes this behaviour to
make OE-based distros behave like other more mainstream distros.

Removing the global ulimit calls is scheduled for the 2018
release (it's too late to do this in 2017 because we risk invalidating
manual tests) so this is just a temporary workaround until 2018.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>