packagegroup-ni-base: install dmidecode only for x64
Upstream salt alawys assumes that if dmidecode is present the system
is running a BIOS and tries to read the DMI tables using these tools.
We want to add logic to salt to read uboot env vars via fw_printenv
and we don't want to break salt upstream's assumption.

Also dmidecode is completely useless on ARM so it makes sense to
remove it from our safemodes/runmodes because ARM tragets are
more memory constrained and we save around 100kb of space.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>