zynq-bootscripts: Add SET_HW_ENV_INFO
Salt queries U-BOOT env vars for it's ARM hwinfo grains. For now we
just need 'serial#' and 'manufacturer'. serial# is already set within
uboot before the bootscripts are run. manufacturer and possibly others
are not, so we need to add them in the bootscripts to make ARM salt
work for already produced devices which don't expose 'manufacturer'.
This env vars list may grow in the future.

This works perfectly because NI does not currently support any
non-NI ARM hardware on NILRT. If/when we will support non-NI ARM
targets we'll have to set the env vars contained in SET_HW_ENV_INFO
conditionally (maybe check some cpld register to determine the
hardware is really made by NI?).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>