busybox: allow zcip action script to reject chosen IP
There may be valid reasons to reject a chosen IP for reasons beyond what
can be determined by watching ARP updates on the specified interface. In
particular, this LLA could have been chosen on another interface.

Traditional solutions to this problem involve listening to
ETH_P_ALL (slow) or libpcap (big). Instead, as a first-order workaround,
delegate the complexity of this problem to the action script, by letting
it reject the chosen IP by returning error 67 (ASCII 'C'). It is assumed
that the action script has left the interface deconfigured.

Not recommended to upstream this because, technically, it's not a bug.

Natinst-CAR-ID: 626575
Natinst-Reviewboard-ID: 171547
Acked-by: Eric Gross <eric.gross@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>