refpolicy: Add bbappend files for ni-refpolicy
Add bbappend files to use the ni-refpolicy.  Also update the
include file to use the krogoth branch and the meta-conf layer.

Leaving the old bbappends as part of this commit, because upstream
is continuing to provide recipes for both versions.  So to continue
to be compatible with upstream, the files were copied instead of
renamed.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>