ninextos: depend on restore-mode-image's do_image_complete
The necessary file [1] for inclusion in ninextos is not created at the
end of do_rootfs and this causes build failures and nasty hacks in
jenkins to build the ninextos package in a separate bitbake call than
the restore-mode-image.

By depending on do_image_complete we make sure the cpio.gz is present
and avoid the build failures & jenkins workarounds.

[1] tmp-glibc/deploy/images/x64/restore-mode-image-x64.cpio.gz

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>