xfce4-panel: fix custom embedded-ui icons
A change in 2016-2017 krogoth OE upgrade broke these icons in the
Embedded UI because upstream xfce change the path from
/usr/share/icons/*x*/* to /usr/share/icons/hicolor/*x*/apps/*

Also installing a custom icon theme in the do_unpack python function
is OE non-standard, instead we put them in do_install (additionally
for compiled sources they should be added to do_patch & do_compile in
addition to do_install, but this is not the case here for the icons).

Also archive the icons to reduce their size from ~60kb to ~12kb.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>