nilrt.conf: split & add older nilrt support
We now support 3 distributions on the same branch:
  1. Newer NILRT (nilrt-nxg.conf)
  2. Older NILRT without UI (nilrt.conf)
  3. Older NILRT with XFCE UI (nilrt-xfce.conf)

The common logic for all these distro configs is in nilrt.inc.

The majority of logic between these distros is identical and there are
only a few components whoose recipes get separated in the meta-nilrt
layer based on the distro configuration used at build.

Maybe in the future we can do some more work to reduce the abstraction
burden between nilrt and nilrt-xfce (2 and 3); it shouldn't be hard to
merge them if we prove that enabling the x11 DISTRO_FEATURE in the
non-UI nilrt distro doesn't increase size footprint too much.

Maybe a few years ago, with the process, knowledge and OE state at the
time, keeping distros 2. and 3. separate made sense, but we should
re-analyze that decision and merge them if possible in the future to
reduce the maintenence burden.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>