x64.conf: move MACHINE_ESSENTIAL_EXTRA_RRECOMMENDS pkggroups
In meta-nilrt this variable is only used for x64, not for arm.

First of all, this variable is misused. According to it's doc " (...)
Because this is a 'machine essential' variable, the list of packages
are essential for the machine to boot." It makes no sense to install
stuff like wacom tablet & touchscreen drivers in it. None of the
packages installed here are essential for booting (all that is built
directly in our kernel I assume, even e1000e is not essential for
booting even though networking won't work) so we can remove them all.

Secondly, this is the _only_ difference between the older and newer
nilrt x64 machine definitions. If we want to keep this only for the
newer NILRT's then we'll have to fork the machine definitions (we
already have them forked now on different git branches and we want
to unify them, it makes no sense to keep separate x64 definitions for
exactly the same hardware machines).

So just install these packages directly in the base packagegroup like
the rest of the packages since the MACHINE_ESSENTIAL_EXTRA_RRECOMMENDS
variable was used only there and avoid the need for 2 separate x64
machine definitions for the older/newer NILRT's.

In the future I'd recommend we remove as many unnecessary packages
from this list to save space on disk (do we really ship targets with
wacom tablets? or run vmware vm's on the devices themselves? etc).
Maybe we can put some of these in the feeds?

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>