bash.bbappend: source /etc/profile in non-interactive shells
Follow the standard way all distributions sounce /etc/profile (OE
upstream started doing it in their krogoth release) now that the
stack limit was increased to 512kb so the few offenders (thunar, etc)
don't fail anymore due to hitting the limit (previously was 256k).

This was requested by the salt-based teams running on NILRT so they
can have the same environment set up in non-interactive shells as
when they run salt locally in an interactive shell.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>