local.conf.sample: Remove ASSUME_PROVIDED sdl-native
In the past OE was unable to build libsdl, but is now able to.  In order to
workaround this libsdl had to be installed on the host system and
ASSUME_PROVIDED was used to force OE to use it.

This commit removes the ASSUME_PROVIDED, just as it was removed from
openembedded-core/meta/conf/local.conf.sample, to force OE to build libsdl.

Note that the ASSUME_PROVIDED was causing a build failure on Ubuntu 16.04
based systems since OE was unable to find the host's libsdl-dev package.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>