packagegroup-ni-extra: Fix nilrt (non-xfce) build
When merging older and newer meta-nilrt branches I forgot to protect
the GUI packages under a DISTRO_FEATURES containing an x11 check so
the Jenkins feed builds started failing for older non-UI NILRT because
it has x11 disabled (of all the local build combitations which I did
for nilrt/nilrt-xfce/nilrt-nxg + x64/arm + base/images/extras, I
missed this one, sorry).

I really really hope in the future we can avoid biffurcating the older
distro so we don't have to make a special case to not include the GUI
components in its feed (I see no real reason to have the biffurcation,
but since it's the status quo we need to do some thorough disk space
analysis before removing it to make sure we're not breaking anything).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>