opkg: Pull source from git repo
We require support for verbose_status_file and a fix for spaces in user-defined
fields that are present in the master branch of opkg, but those changes are not
part of an opkg release yet.  For now pull from the latest commit of the master
branch.

This will also allow more testing for the opkg master branch, which should make
future opkg version upgrades smoother.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>
Natinst-CAR-ID: 641602,632922