grub.cfg: Remove consoleparam override on non-NI hardware
The default consoleparam of "console=tty0 console=ttyS0,115200n8" works
better on VMs because it directs the kernel to display the console on
both video and the first serial device. Overriding it to "console=ttyS0"
on non-NI hardware disables the console for most users, which use a
default IO configuration on their hypervisors (video only, no serial).

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-ReviewBoard-ID: 185353