lvcomms-nilrt-image: Remove ni-pxiplatformservices
With the recent opkg upgrade, the image recipe is unable to build due to some
new errors generated by the PXI packages. Specifically, files are being
installed under some symlinked directories which is disallowed in newer opkg
versions.

Temporarily remove the PXI packages to allow time for them to be fixed.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>