n310: Add machine conf
This machine configuration is meant to house the differences
between normal NI Zynq target machine configurations and those
of the NI/Ettus N310. Examples of these differences:

> SD card for boot medium
> u-boot configuration

Ideally, we can expand this notion to have a common .inc that
houses the shared attributes of all Zynq targets and forces
target-specific elections to live in separate .conf files.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>