u-boot-2017.%.bbappend: Add
Adding this bbappend to redirect to our internal git server where
the needed changes to support the N310 have been made on an internal
branch. Mostly, this consists of a defconfig and a ps7 .c (for SPL/
FSBL reasons), but this also allows us to pull in any ChromeOS or
DS1374 WDT changes for supporting the target.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>