apr-iconv: allow building default OE packages
Since we build only the 'apr-iconv-src' ipk but specify the dep on
'apr-iconv' inside the packagegroup, dependencies are auto-added by
opkg for implicit default packages ${PN} ${PN}-dbg etc. which do not
exist and then the opkg-graph-deps tool which tests the Package index
during jenkins builds finds an error.

The simplest solution is to depend directly on apr-iconv-src in the
packagegroup but this is a little messy because that exact name from
the packagegroup is also used to add dependencies in the index file:

Package: packagegroup-ni-coreimagerepo-dbg
Version: 1.0-r0.3
Recommends: (...), apr-iconv-src-dbg, (...)

which also doesn't exist and will trigger other test errors.

The best solution IMO to fix all these broken dependencies is to build
the default apr-iconv packages and let them stay in the feed just like
we also do for apache2.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>