nilrt.conf: Alias n310 feed URI to match xilinx-zynqhf
Without this change, the URI would have duplicated the otherwise-
identical feed structure under a .../$MACHINE/$SUBFEED structure
(which contains folders for $TUNE_PKGARCH, $MACHINE, and all, only
the $MACHINE folder will differ).

With this change, we'll still have the separate $MACHINE subfeed
but the identical $TUNE_PKGARCH and all folders will not be duplicated.

Ideally, the TUNE_PKGARCH would be used as the designator, but we've
sadly built tool-based dependencies on this mapping.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>