n310: Move U-Boot variables into machine conf
Initially, the approach was going to be to include additional
OVERRIDES when building for n310 (without the use of a separate
MACHINE). When this proved to be too problematic an approach,
a separate MACHINE was defined but the things that should exist
within that MACHINE definition never made their way back to
the config. Move the U-Boot MACHINE-specific variables to the
MACHINE configuration for the n310.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>