n310.conf: Workaround wic double-build failure
Currently, when building a wic image, bitbake will attempt to
determine the size of the resulting image (as reported by
ROOTFS_SIZE), which is part of do_rootfs_wicenv[vardeps]. This leads
to do_rootfs_wicenv[prefuncs] containing "set_image_size", which
will update the current ROOTFS_SIZE variable by calling "du ..." on
the IMAGE_ROOTFS directory. Doing a second build that does not have
changes that impact the rootfs for the image will lead to the hash
being the same for the IMAGE_BASENAME, meaning that it will not
repopulate the rootfs, and this check (which is "needed" to determine
whether or not we need to re-wic-ify) will attempt to check the
rootfs size of a folder that got cleaned up on completion.

All of this is to say, until the issue is resolved, workaround
the issue by keeping around the rootfs for the image.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>