fw-printenv: fix variable quoting
Using double quotes is unnecessary and splits the string in many
pieces which get concatenated at runtime with the variables. The most
annoying thing about this is the multi-color syntax highlighting it
introduces for the string / variable / string / variable succesion.

Just use a single string for everything because bitbake is perfectly
capable of variable interpolation after calling the function.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>