nilrt.inc: bump GCC 5.3 -> 6.4
OE Pyro only supports GCC 5.4 or newer 6.3. An increasing number of
recipes are broken under GCC 5 (for example tbb assumes -flifetime-dse
which is only present in GCC >=6). If we want to stay with GCC 5 we'll
have to fix these upstream failures ourselves.

Also GCC 7 is out. OE upstream will drop GCC 5 in the next OE release
post-Pyro so we might as well uprev now.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>