gcc*.bbappend: bump gcc 4.% -> %
These bbappends slipped last year when we upgraded gcc 4 to 5 and
no-one encountered the bug these were fixing because (I assume) no
one compiled & ran software directly on ARM targets for the past
year to test their floating point performance. On-target compiled
binaries still work correctly without this change, it's only that
floating point operations should slower without it.

The correct fix to avoid these kinds of problems in the future
(we'll be updating to gcc 6 next rebase because Pyro is the last
release OE upstream supports gcc 5) is to make the bbappends
versions a full wildcard % because the code in them is very
generic and only depends on the presence of a few flags like
'--with-arch=' which don't really change between GCC versions.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>