refpolicy: bump revision and use git recipe
NI's refpolicy was using a non-git recipe for a git repo build so even
though the non-git recipe was uprev by upstream to 2.20151208 and we
bbappended it, we overwrote the SRC_URI and used an older
version (2.20140311) even though the ipk's will have the upstream
version.

Also make the git revision autorev like all other of our git recipes
so we don't have to hardcode each revision in the recipe file.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>