linux-nilrt: Add kernel tagging task
With the introduction of a means to set the kernel's LOCALVERSION,
the next logical step is to tag the kernel source (if requested).
This will make matching a booted kernel to a particular commit easy,
and, with a LOCALVERSION scheme that changes predictably (i.e.
increments on successful builds), it is simple to tell whether or not
a booted kernel contains a particular commit.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>