linux-nilrt.inc: add DEPENDS on rsync-native
Recipe specific sysroots were added to OE-core and this means that if
a dependency is not explicitly specified it will not be found in the
recipe's sysroot (as opposed to using a common sysroot for all
recipes). So now we need to explicitely add all tools we depend on
otherwise our build will fail because they're not found.

Also fix a variable overwriting bug in the arm dtc-native depends.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>