recipes-*: fix useradd sysroot dependencies
Because of the recipe specific sysroots feature added in Pyro,
dependencies do not get automatically added to a recipe's sysroot
during build unless explicitely specified (this happened implicitely
before because builds used a common sysroot) to avoid build races and
improve reproducibility.

In the case of recipes depending on niacctbase we have to also
explicitely depend on shadow-native and pseudo-native otherwise the
users won't be added in the dependent recipes sysroots and they will
error with user/group not found.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>