lvcomms-nilrt-image: Organize installed NI software
Since NXG does not mean x64-only, cleanup and organize the packages
installed into the image to reflect this fact.

Basically, move the clearly-x64-specific components (centered largely
around working with PXI devices) to the x64-specific append. This
means trimming down the MACHINE-agnostic list to only include
packages that should exist on all MACHINEs.

For packages that are not yet ready for armhf, keep the packages in
the MACHINE-agnostic list but remove them in the armhf-specific
_remove list.

While working on this, a duplication was noted with the salt minion
package (being in both the external-components.inc as well as the
image definition), and the underlying package has changed names
(niminionagent->ni-sysmgmt-salt-minion-support), so rename this
package and move to a MACHINE-agnostic _append.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Natinst-ReviewBoard-ID: 195385