restore-mode-image: fix $FACTORY_IMAGE dependency
Depend on do_image_complete instead of do_tar to avoid a build race
because we don't know if the tarball got copied to deploy/images after
do_tar (we only know that the tar was built). By depending on
do_image_complete we are assured the $FACTORY_IMAGE build is 100% done.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>