u-boot: add fw_env-xilinx-zynq.config
This change was tested only on NXG branch nilrt/cardassia/krogoth and
it was not tested on the older nilrt branch with the plain
xilinx-zynq (not hf) machine.

Because we're on the merged meta-nilrt branch we need to provide this
file to avoid a warning and error while building u-boot for CG (the
hardware and mem locations are the same as with the hardfloat build).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>