migrate-nilrt: migration with force safemode enabled
When migrating a target from CG to NG which have force safemode enabled,
the target is still booting in safemode after running ni_migrate_target
and will continue the migration only at the next normal boot(runmode).

This change makes sure that the next boot will be runmode, regardless if
force safemode is enabled or not.

Natinst-ReviewBoard-ID: 198166
Signed-off-by: Heghedus Razvan <razvan.heghedus@ni.com>