packagegroup-ni-coreimagerepo: don't build xfce & xserver on nilrt
We don't build these in the older 'nilrt' distro in our jenkins build
job (that's why this hasn't failed any build, we use the nilrt-xfce
distro for this) but we should prevent anyone from manually building
this packagegroup with xfce and xserver included in 'nlirt' because
it's not supported.

For building xfce/xserver in the older distro, please use nilrt-xfce.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>