coreutils: fix coreutils-native build
We don't build nor care about hostname/ls/chcon in coreutils-native
but the rdepends was added for both target and native so the native
build was broken until now.

We didn't catch the broken coreutils-native because nothing we
shippend depended on it until openvswitch in the pyro extras feed.

Since the fix is trivial I'm opting to fix instead of removing
openvswitch from extras.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>