grub-efi: bbappend oe-core recipe
meta-measured created the Pyro branch in which it removed the grub-efi
recipe we were bbappending:

commit 68e6bb6f29340720bbdf4ca2d001b498e3bded70
Author: Philip Tricca <flihp@twobit.us>
Date:   Fri Jul 28 01:53:05 2017 +0000

    grub-efi: Remove TPM2 patches.

    Patches from mjg59 have landed on the grub mailing list. I don't plan to
    maintain these patches. Future work should be based on the work going
    upstream.

    Signed-off-by: Philip Tricca <flihp@twobit.us>

Bbappend the oe-core recipe even if it's not a git recipe (there's no
grub-efi_git recipe anywhere now). We don't have to do anymore logic
changes because we're bundling our logic in our own grub-efi-nilrt
package.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>