busybox: make recipe machine specific
There is a bug in the busybox recipes triggering rebuilds between
xilix-zynqhf and n310 for busybox and busybox-udhcpc which bump the PR
number triggering build failures in the jenkins pipelines.

I suspect the cause it that the MACHINE variable invalidates the
busybox-udhcpc sstate, but I couldn't root cause it and I have no more
time to investigate so just make busybox MACHINE-specific until more
time is found to fix it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>