busybox: make recipe machine specific
On the krogoth branches there is a bug triggering rebuilds between
xilix-zynqhf and n310 (it doesn't reproduce on pyro) which break the
build when building in place inside the NXG pipeline. Root causing and
fixing this is hard and we're about to switch to pyro anyway so just
make busybox machine specific to avoid xilinx-zynqhf and n310
clobbering eachother.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>