distro-feed-configs: Genericize one-off URIs
Previously, the migration feed URI was unique in that it was a
subfeed that only defined one (or, really, less than every possible)
Architecture and therefore was ill-suited for the templatized URI
handling in the distro-feed-configs addition. This was OK for a one-
off feed, but there will likely be others to follow and this scheme
doesn't allow for easy additions.

Here, the special URIs are defined in the same manner as other feed/
URI tuples in OE to allow for simpler addition to feeds that, for one
reason or another, don't fit into the templatized scheme defined by
the permutation of NILRT_SUBFEED_NAMES and NILRT_SUBFEED_ARCHES.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>