init-runmode-ramfs.sh: Add encrypted disk support via nilrtdiskcrypt
If a default tpm is present (at /dev/tpm0), the init-runmode-ramfs.sh
will now reseal open keys to current PCR state, then nilrtdiskcrypt_open
the specified partition and boot it.

Additionally, this change improved diagnostic output of the script when
the "debug" kernel parameter is present. Prints output to kmesg buffer,
so that it's readable after boot.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-Rally-ID: F12416
Natinst-ReviewBoard-ID: 204022