opkg_%.bbappend: Remove faulty tests
As noted in CAR 670541, some of the opkg tests have faults in how
they test (try to install cURL to verify that the feeds work when
cURL already exists on the target), and both removed tests modify
the target unneccesarily. Until such a time as the tests are fixed,
remove them from being installed.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Natinst-CAR-ID: 670541
Natinst-ReviewBoard-ID: 205163