nilrtdiskcrypt_format: prevent formatting if PCR11 is set
If key unsealing is disabled (PCR11 is non 0), don't allow formatting
with encryption because it will render the system unbootable (the
value of PCR 11 is lost at reboot by design so we don't want to seal).

It also makes logical sense: if unsealing is disabled, any kind of
system updates affecting encryption are disabled, including format.

What can only be done is a full format without encryption, in this
case the state of PCR registers doesn't matter and all data is lost
anyway.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
Natinst-Rally-ID: F12416
Natinst-ReviewBoard-ID: 205398