nilrtdiskcrypt_unseal: add source/dest arguments
We want to provide a way to unseal a safemode key slot and store the
unsealed key in both the safemode and runmode key slots at the same
time (both unsealed and in runmode's case it will seal it's key at
next boot), overwriting any existing key value.

To do this we allow specifying one source slot and multiple
destination slots.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>