packagegroup-ni-base: remove util-linux-findfs
This tool was used to find fs labels in the nisystemformat script, but
once we implemented TPM-backed encryption format support we stopped
using it because it doesn't work with luks containers and replaced
calls to it with 'lsblk -o LABEL/PARTLABEL' calls.

Remove it to free up some disk space because nobody is using it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>