nilrt-image-common.inc: Move INITRAMFS_FSTYPES to xilinx-zynqhf
This variable is machine specific by definition and can't be overrided
using the machine variable, it needs to live alonside the rest of the
machine conf and the IMAGE_CLASSES += "image_type_uboot" declaration.

Without this the build generates normal cpio.gz not cpio.gz-u-boot
which will not be bootable.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>