packagegroup-ni-runmode: remove ni-module-versioning-tools
This recipe was was added for NXG but was never used in the runmode
packagegroup and after I merged the meta-nilrt branches I added it
to the older nilrt runmode also increased its size by over 100 mb
because it pulls gcc and other packages as dependencies.

Remove it to reclaim space, it doesn't affect functionality because it
was never in the image to begin with.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>