nilrt.inc: remove dangerous useless VIRTUAL-RUNTIME_dev_manager
Not only this line is useless (it's only useful if we switch to
busybox-mdev, grep the sources for it) it's actively misleading
making you think we're actually using systemd's udev when in reality
we're using eudev because we pull udev-extraconf which rdepends on
eudev in packagegroup-ni-base (systemd is disabled in DISTRO_FEATURES).

Also explicitely add eudev to the base packagegroup because it's
implicitely pulled in by udev-extraconf which is very ugly.

Also there is no more "udev" in OE, if we ever want to go back to it
we need to include systemd which provides udev (udev got merged into
systemd).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>