tpm2-abrmd: fix build
This is a deprecated component we're still using and which we'll
remove once we bump the linux kernel past 4.12 which contains an
in-kernel tpm resource manager.

Nobody cares about this component and upstream discourages its use so
fix it here until we bump the kernel. meta-measured upstream will
eventually remove it also.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>