linux-nilrt: Improve latency by disabling the i915 video driver power-saving
A 300+ uS latency spike was observed in 4.9-rt and later kernels when render
and display power-saving states are enabled in the Intel i915 GPU driver
(default in 4.9).

Our general policy is to disable power management in order to improve
real-time performance. Apply that policy to the i915 driver by setting:
  * enable_rc6=0 - disables power-saving render C-state 6;
  * enable_dc=0 - disables power-saving display C-states;
via the module_conf_* mechanism which generates the '/etc/modprobe.d/i915.conf'
configuration file.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-CAR-ID: 680530
Natinst-ReviewBoard-ID: 218598