initscripts-nilrt: nivalidatesystem -> nisetled
nivalidatesystem was used in the 2012-era to validate an install at
boot. Since then all the logic was removed from it in p4 except a call
to status_led, then it got blindly copied into git (in older nilrt the
file is still populated from p4, only on nxg it's installed from git).

After the meta-nilrt merge the git file got included in older NILRT
triggering CAR 684261.

So... rename the initscript and install only in NXG.

Natinst-CAR-ID: 684261

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>