opkg.conf: Move cache to non-volatile memory & add comments
Because on ARM we limit /var/volatile to 64mb and on x64 it's limited
to half the device RAM size, installation of big packages (or ipks
having lots of dependencies) exceeding the limit will cause opkg to error.

Move the cache dir to non-volatile memory so software installation is
not bounded by the device RAM size.

Also since I'm at it, add comments before the config options to make
it clear what's happening with each of them (useful for clients adding
their own feed etc).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>