nilrt.inc: clenup & fix glibc binary compatibility
glibc uses the $OLDEST_KERNEL variable to determine which ABI to use
or which features / optimizations / new syscalls are available to make
its binaries compatible with kernels as old as $OLDEST_KERNEL.

On ARM we depend on specific optimizations and fixes from newer
kernels to avoid problems like bug [1] which  effectively disables
priority inheritence, so we set it to 3.14.3.

On x64 we also shipped with 3.14.3 last year and it's safer to not
break compatibility, especially this late in the release cycle.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=18463

Natinst-CAR-ID: 684246

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>