minimal-nilrt-image: Add license tarball to image
In the fido branch, we added support for compressing the license files
produced by the COPY_LICENSE_MANIFEST feature and include the compressed
tarball in the image. commit 6db398341fea4a1d0248ec1a51c81489ffa7af3e

This approach required changes to oe-core's license.bbclass to allow it to
work with BUILD_IMAGES_FROM_FEEDS.

This commit adds the license tarball to the image but does it in a way that
uses LICENSE_CREATE_PACKAGE, and doesn't require modification of the
license.bbclass.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>
Natinst-ReviewBoard-ID: 222476